Make `CounterType` non-generic
